:truck: Move BasePolicyScope to its own file.
Why? One class per file scales better.
