+++ src/api/policies/base-policy.ts
@@ -1,24 +1,4 @@
-import knex, { Knex } from 'knex';
-
 import { User } from '../models';
-
-export abstract class BasePolicyScope {
-	scope: Knex.QueryBuilder;
-	user: User;
-
-	constructor(scope: Knex.QueryBuilder, user: User) {
-		this.scope = scope;
-		this.user = user;
-	}
-
-	resolve(): Knex.QueryBuilder {
-		throw new Error('not implemented');
-	}
-
-	get emptyScope(): Knex.QueryBuilder {
-		return this.scope.whereRaw('(1=0)');
-	}
-}
 
 export abstract class BasePolicy<Type> {
 	record: Type;
