+++ src/api/policies/index.ts
@@ -1,3 +1,4 @@
 export * from './base-policy';
+export * from './base-policy-scope';
 export * from './place-policy';
 export * from './place-policy-scope';
